+++ CHANGELOG.md
@@ -16,6 +16,7 @@ All notable changes to this project will be documented in this file.
 - New `noop` cache.
 - Field `targets_input` added to the `azure_blob_storage` input.
 - New `reject_errored` output.
+- New `nats_request_reply` processor.
 
 ### Fixed